refactor: [FE] 결과 스케치북의 우측 컴포넌트 리팩터링
- 결과스케치북 우측에 들어가는 스케치북 페이지 넘버와 페이지 작성자 이름이 별도 컴포넌트로 분리되었어요.
- ResultSketchbook에 들어가는 컴포넌트를 따로 모아놓기 위해서 하위 폴더를 만들었어요.
